A #using_javascript_driver test helper
When we `require 'slimmer/test'` we override the method that results in a network request to static for the template to stuff everything inside `<div id="wrapper">` into and instead uses a template fixture from slimmer eg. wrapper.html or header_footer_only.html.

These template fixtures include javascript files that are hosted on assets.digital.cabinet-office.gov.uk and therefore require a network request.

Some of the apps including this one, to avoid making requests over the network, have used the following snippet to tell slimmer to not use slimmer unless specified with an environment variable.

```
class ActionController::Base
  before_filter proc {
    response.headers[Slimmer::Headers::SKIP_HEADER] = "true" unless ENV["USE_SLIMMER"]
  }
end
```

Why isn't this snippet always a good idea? When tests need javascript.

jQuery, govuk-template.js and static's application.js are included via the slimmer template. As it stands if your test for example needs jQuery it won't be available unless you:

a) Allow the test to reach a running instance of static by removing `require 'slimmer/text'`.
b) Disable the above snippet so wrapper.html can be used and the externally hosted javascript files pulled in.

This commit is an implementation of b).