Add a test for the existing search header
The existing search box in the service manual header is provided by Slimmer. We set a response header to tell Slimmer to add a hidden field to scope the search results to the manual.

We should test this functionality to ensure it’s not broken when we remove the search box on the homepage.