Add design principles pages.
Again, these are copied more or less directly from the designprinciples
app.

The accessiblepdfs page is currently rendering nav links in the form
design-principles/mainstream?action=styleguide&controller=root -
this appears to be a bug introduced when the other pages were removed
and redirected to the content guidance. Since those links are
redirecting anyway, in porting them over they have been hard-coded to
point to the guidance page itself.