Switch to new Design System link styles
The new DS link styles are being rolled out across GOVUK.
We've already applied the new styles to the individual components in the
component gem. This switches service-manual-frontend over to the new
styles by setting the govuk-new-link-styles flag to true in the
application.scss file.

Additionally: replace instances of app-link--inverse with
govuk-link--inverse and remove redundant app-link--inverse CSS