Fix only serving service_manual_ document types
This removes a bogus test that checked the homepage returned a
particular error. This test didn't actually work for the scenario of
Service Manual trying to render the real GOV.UK homepage. This was
because the mechanism to determine a presenter used the removal of a
prefix from a string (such as service_manual_homepage) so when the
actual homepage came along, with it's document_type of "homepage"
there'd be no prefix to remove and the page would render.

This updates the code to actually raise on any document types that don't
have a prefix of service manual.