Update README for current GOV.UK conventions
This updates this readme to be consistent with other GOV.UK applications
that use GOV.UK Docker.

As this application doesn't get much love I've been quite brutal in
removing things that are unusual for a GOV.UK frontend application.

I've removed the information about not serving the homepage as I believe
this is sufficiently covered by the error message in the previous
commit. This behaviour is also consistent across virtually all other
GOV.UK frontend applications and not documented in this way.

I've removed the content schema information too as this is
unconventional. When I tested this repo without content-schemas
available I noticed that you get essentially the same guidance in an
exception:

```
Dependency govuk-content-schemas cannot be found at: /Users/kevindew/govuk/govuk-content-schemas. Clone it to that directory, or set GOVUK_CONTENT_SCHEMAS_PATH (see README.md for details). (GovukContentSchemaTestHelpers::ImproperlyConfiguredError)
```