Remove unnecessary Dockerfile
This application is not part of the publishing-e2e-test suite so there
is nothing that actually makes use of this Dockerfile. It is also
incomplete as it lacks the building of assets.

To complete this work the Docker image will need to be removed from
Dockerhub otherwise deployments will be flagged as unstable (due to
Dockerhub no longer being up-to-date). This should have no practical
consequence as nothing should be using these Dockerhub images.