Remove specified search id
Previously, clicking on the label in the search box did not move focus to the search input, which made the search box feel broken.

This was happening because there are multiple things on the page with an id of 'search'. The search component generates an id with a unique number appended to avoid this happening. There doesn't seem to be a reason why this input needs a specific id, so this commit removes it so it gets generated instead.